Address possible loss of data warnings
From Visual Studio, for example:

_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(160): warning C4018: '<': signed/unsigned mismatch
flag
Build Log Line	24
Warning
_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(316): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
flag
Build Log Line	25
Warning
_deps\cleaver_lib-src\src\lib\cleaver\CleaverMesher.cpp(433): warning
C4244: 'return': conversion from 'double' to 'float', possible loss of
data

+++ src/lib/cleaver/SizingFieldCreator.cpp
@@ -99,17 +99,17 @@ namespace cleaver
     return dist[l][m][n];
   }
 
-  int VoxelMesh::distSizeX() const
+  size_t VoxelMesh::distSizeX() const
   {
     return dist.size();
   }
 
-  int VoxelMesh::distSizeY() const
+  size_t VoxelMesh::distSizeY() const
   {
     return dist[0].size();
   }
 
-  int VoxelMesh::distSizeZ() const
+  size_t VoxelMesh::distSizeZ() const
   {
     return dist[0][0].size();
   }
